volcartesian: fix m_directionOrdering declaration
m_directionOrdering is declared as a 3-array of doubles
but used as an integer array.

+++ src/volcartesian/volcartesian.hpp
@@ -169,7 +169,7 @@ private:
 	std::array<double, 3> m_minCoords;
 	std::array<double, 3> m_maxCoords;
 
-	std::array<double, 3> m_directionOrdering;
+	std::array<int, 3> m_directionOrdering;
 
 	std::array<std::vector<double>, 3> m_vertexCoords;
 	std::array<std::vector<double>, 3> m_cellCenters;
